chore(sortable-table): move types to types folder
