+++ packages/blockchain-wallet-v4-frontend/src/scenes/Interest/EarnTable/SortableTable/SortableTable.types.ts
@@ -1,0 +1,12 @@
+import { ButtonCellProps, TextCellProps } from '@blockchain-com/constellation'
+
+export type RowType = {
+  actions: ButtonCellProps | TextCellProps
+  asset: TextCellProps
+  balance: TextCellProps
+  rates: TextCellProps
+}
+
+export type CellProps = {
+  value?: number | string
+} & TextCellProps
